converting bytes to uint16 as defined by business rules

+++ App/Program.cs
@@ -9,7 +9,7 @@ namespace App
             var client = new Client("40.121.85.141", 20010);
             client.Connect();
             Console.WriteLine(client.ReadSerialNumber());
-            int[] registers = client.ReadRegistryStatus();
+            ushort[] registers = client.ReadRegistryStatus();
             Console.WriteLine(registers[0]);
             Console.WriteLine(registers[1]);
         }
